third_party/blink/web_tests/external/wpt/dom/events/webkit-transition-end-event-expected.txt

This is a testharness.js-based test.
FAIL ontransitionend and onwebkittransitionend are not aliases assert_equals: onwebkittransitionend should initially be null expected (object) null but got (undefined) undefined
FAIL dispatchEvent of a webkitTransitionEnd event does trigger a prefixed event handler or listener assert_equals: onwebkittransitionend must be supported and not previously set expected (object) null but got (undefined) undefined
PASS dispatchEvent of a webkitTransitionEnd event does not trigger an unprefixed event handler or listener
FAIL dispatchEvent of an transitionend event does not trigger a prefixed event handler or listener assert_equals: onwebkittransitionend must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkittransitionend event handler should trigger for an animation assert_equals: onwebkittransitionend must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkittransitionend event handler should not trigger if an unprefixed event handler also exists assert_equals: onwebkittransitionend must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkittransitionend event handler should not trigger if an unprefixed listener also exists assert_equals: onwebkittransitionend must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL event types for prefixed and unprefixed transitionend event handlers should be named appropriately assert_equals: onwebkittransitionend must be supported and not previously set expected (object) null but got (undefined) undefined
PASS webkitTransitionEnd event listener should trigger for an animation
PASS webkitTransitionEnd event listener should not trigger if an unprefixed listener also exists
PASS webkitTransitionEnd event listener should not trigger if an unprefixed event handler also exists
PASS event types for prefixed and unprefixed transitionend event listeners should be named appropriately
PASS webkitTransitionEnd event listener is case sensitive
Harness: the test ran to completion.
